Updated versions.
Description of the Change:

Updated the version 1.8 of JDK
Updated dependencies to latest releases.

Benefits:
Will allow to be compiled with latest JDK as 6 and 7 are no longer
supported.
Will allow to use JDK 8 features such as streams
Less bugs in later releases (in theory)

Possible Drawbacks:
Legacy application using jdk 6 will not longer work with this version,
hoever JDK 6 was EOL in 2018.

Applicable Issues:
None.